DataEngineering/Airflow/requirements.txt
# -*- coding: utf-8 -*-
--index-url https://pypi.org/simple
apache-airflow~=2.0.0
